Remove blank keys
With the spreadsheet setup below, I was receiving dictionaries with an extra key/value pair in which both the key and value were an empty string.  Proposed changes will skip the iteration if the key is a blank string.  I thought about checking if the value was an empty string, but the end user may want that behavior.