Inherent Types and Arbitrary Properties
Instead of having a column for type that applies to a property column, this update makes the property have the type inside itself. Much better UX and smaller sheets.

Once we no longer need a typing column, then we don't need to enforce a specific sheet structure. This update includes the ability to put any amount of arbitrary properties per Value.

This update also moves the type transformers into a separate file for maintainability and adds a few more types.